Massive changes, whoops.
Project files are now pretty-printed; diffs are easier to see.

Adding an "assets/icons.png" file for different small icons.

Color themes are better implemented; now they are put into a map instead of directly being held in variables. Those variables are now constants for easy reference to the different colors used in the palette for the theme. For this, I also added getThemeColor() to easily retrieve the color for rendering.

Color themes are now saved in project save file.

Adding search bar to search for Tasks by words in the description.

Adding method to focus view on selected Tasks.

Adding project.MousingOver() to easily tell what "zone" of the screen the mouse is in.

Project saving now happens when a message is sent (so generally whenever anything important happens), or when panning or zooming. This should assist with keypresses that weren't letters or numbers not saving.

Comma and Period now walk through Tasks sequentially.

Directional keys now move up and town a task tree.

Shift and directional keys now rearrange a task in a task tree, or indent / outdent accordingly.

Creating a new Task now inserts it above the currently selected Task.

Deleting a Task now rearranges the tree below it, moving all Tasks upwards.

GenerateGrid is now in the project struct, not a free-standing function.

Tasks now can have vertical neighbors.

Tasks are now automatically numbered if they are touching vertically. This will depend on the project's reordering settings (it can be turned off).

Images are resizeable; when selected, they will display two squares, allowing you to resize or reset the size.

Sound file groundwork is being laid.

Fix: Can click on status bar and have Tasks register it.

Fix: Can click on timescale bar and have Tasks register it.

Fix: Failure to load a file won't crash, but rather print a line to log and return.

Moving NewProject to top, under struct declaration.