Improve HOME/END text selection
With the previous behaviour, if one placed the caret in the middle of a
text selection, held shift, pressed END and then HOME, the selection
would look like this:
|hello world| (the whole text box is selected)
while a "normal" text box would act as if END was never pressed to begin
with. This commit changes the behavior to act like the latter.