QoL: Adding Web browser card. QoL: Fixing card shading to be consistent, and not for unshadeable cards (map and table, for example). QoL: Adding time mode to Timer card. When in time mode, the card will enagage whenever the time is past the time on the card. When a checkbox is above it and enabled, the card is engaged, and if it is above and not enabled, the card is not engaged. This can be used to make everyday, regular todo lists. QoL: Adding keybinding to reset card size to original size. FIX: GUI elements now properly color according to the theme (buttons sometimes were miscolored). QoL: Pushing map elements now loops the elements around the map.
QoL: Adding options to rotate and flip maps in the map editing menu.
Minor QoL: Scrollbars that appears only when the mouse is near now are a little less sensitive.
Minor QoL: It's now easier to scroll things to the top or bottom (by adding a margin).
QoL: Adding option to reset all settings in General Settings section.